pyproject.toml
[tool.poetry]
name = "madapple"
version = "0.1.0"
description = "BMSTU Bachelor's Diploma Fruit Project (aka MadApple)"
authors = ["Valeria Avdeykina"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
# Зависимости PyTorch с конкретными URL для Windows + CUDA 11.8
torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.7.0%2Bcu118-cp310-cp310-win_amd64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu118/torchvision-0.22.0%2Bcu118-cp310-cp310-win_amd64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu118/torchaudio-2.7.0%2Bcu118-cp310-cp310-win_amd64.whl" }
# Остальные зависимости
albumentations = "^1.3.1"
pycocotools = "^2.0.7"
matplotlib = "^3.7.2"
numpy = "^1.24.3"
tqdm = "^4.66.1"
pillow = "^10.0.0"
scikit-learn = "^1.3.0"
opencv-python = "^4.8.0"
networkx = "2.8.8"  # Фиксированная версия для совместимости с PyTorch

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

# Специальные настройки для использования CUDA
[tool.python-system-requirements]
cuda = "11.8"
cudnn = ">=8.0"